# docs/openapi.yaml
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ecotrip internal API",
    "version": "1.2.0",
    "description": "Contrat MVP. Le catalogue, les capabilities et la recherche de trajets utilisent des scénarios synthétiques hors ligne explicitement marqués demo. Aucun compte, réservation, horaire ou offre réelle."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/journeys/search": {
      "post": {
        "operationId": "searchJourneys",
        "summary": "searchJourneys",
        "x-implementation-status": "implemented",
        "responses": {
          "200": {
            "description": "Réponse conforme au contrat",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneysResponse"
                },
                "example": {
                  "dataMode": "demo",
                  "request": {
                    "originId": "demo-paris",
                    "destinationId": "demo-lyon",
                    "departureDate": "2027-01-15",
                    "returnDate": null,
                    "travelers": 1,
                    "modes": [
                      "train",
                      "coach"
                    ]
                  },
                  "outbound": {
                    "status": "complete",
                    "itineraries": [
                      {
                        "id": "outbound-train-via-macon",
                        "direction": "outbound",
                        "requestedDate": "2027-01-15",
                        "dataStatus": "demo",
                        "legs": [
                          {
                            "id": "outbound-train-1",
                            "mode": "train",
                            "subtype": null,
                            "originId": "demo-paris",
                            "destinationId": "demo-macon",
                            "durationMinutes": 95,
                            "waitingMinutes": 10,
                            "distance": {
                              "km": 250,
                              "method": "scenario",
                              "provenance": {
                                "status": "demo",
                                "sourceIds": [
                                  "demo-journey-scenarios"
                                ],
                                "asOf": null,
                                "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                              }
                            },
                            "schedule": {
                              "departureAt": null,
                              "arrivalAt": null,
                              "provenance": {
                                "status": "demo",
                                "sourceIds": [
                                  "demo-journey-scenarios"
                                ],
                                "asOf": null,
                                "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                              }
                            },
                            "provenance": {
                              "status": "demo",
                              "sourceIds": [
                                "demo-journey-scenarios"
                              ],
                              "asOf": null,
                              "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                            }
                          },
                          {
                            "id": "outbound-train-2",
                            "mode": "train",
                            "subtype": null,
                            "originId": "demo-macon",
                            "destinationId": "demo-lyon",
                            "durationMinutes": 70,
                            "waitingMinutes": 25,
                            "distance": {
                              "km": 155,
                              "method": "scenario",
                              "provenance": {
                                "status": "demo",
                                "sourceIds": [
                                  "demo-journey-scenarios"
                                ],
                                "asOf": null,
                                "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                              }
                            },
                            "schedule": {
                              "departureAt": null,
                              "arrivalAt": null,
                              "provenance": {
                                "status": "demo",
                                "sourceIds": [
                                  "demo-journey-scenarios"
                                ],
                                "asOf": null,
                                "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                              }
                            },
                            "provenance": {
                              "status": "demo",
                              "sourceIds": [
                                "demo-journey-scenarios"
                              ],
                              "asOf": null,
                              "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                            }
                          }
                        ],
                        "durationMinutes": 200,
                        "transfers": 1,
                        "emissions": {
                          "status": "demo",
                          "kgCO2ePerTraveler": 8.1,
                          "kgCO2eGroup": 8.1,
                          "coveredDistanceKm": 405,
                          "totalDistanceKm": 405,
                          "comparable": true,
                          "comparisonKey": "carbon-estimation-v1|kgCO2e|kgCO2e/passenger-km|life_cycle|demo|synthetic_test",
                          "methodologyVersion": "carbon-estimation-v1",
                          "assumptions": [],
                          "legs": [
                            {
                              "legId": "outbound-train-1",
                              "status": "demo",
                              "kgCO2ePerTraveler": 5,
                              "kgCO2eGroup": 5,
                              "factorId": "demo-train-factor-v1",
                              "reason": "Calcul de démonstration (provenance des données du trajet)."
                            },
                            {
                              "legId": "outbound-train-2",
                              "status": "demo",
                              "kgCO2ePerTraveler": 3.1,
                              "kgCO2eGroup": 3.1,
                              "factorId": "demo-train-factor-v1",
                              "reason": "Calcul de démonstration (provenance des données du trajet)."
                            }
                          ],
                          "factors": [
                            {
                              "id": "demo-train-factor-v1",
                              "value": 0.02,
                              "unit": "kgCO2e/passenger-km",
                              "mode": "train",
                              "subtype": null,
                              "geography": "FR",
                              "validFrom": "2026-01-01",
                              "validUntil": null,
                              "scope": "life_cycle",
                              "occupancy": null,
                              "version": "journey-demo-v1",
                              "sourceId": "demo-journey-scenarios",
                              "status": "synthetic_test"
                            }
                          ]
                        },
                        "provenance": {
                          "status": "demo",
                          "sourceIds": [
                            "demo-journey-scenarios"
                          ],
                          "asOf": null,
                          "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                        },
                        "warnings": [
                          "Durées, attentes, distances et facteurs entièrement synthétiques; aucun horaire réel."
                        ]
                      },
                      {
                        "id": "outbound-coach-direct",
                        "direction": "outbound",
                        "requestedDate": "2027-01-15",
                        "dataStatus": "demo",
                        "legs": [
                          {
                            "id": "outbound-coach-1",
                            "mode": "coach",
                            "subtype": null,
                            "originId": "demo-paris",
                            "destinationId": "demo-lyon",
                            "durationMinutes": 360,
                            "waitingMinutes": 15,
                            "distance": {
                              "km": 465,
                              "method": "scenario",
                              "provenance": {
                                "status": "demo",
                                "sourceIds": [
                                  "demo-journey-scenarios"
                                ],
                                "asOf": null,
                                "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                              }
                            },
                            "schedule": {
                              "departureAt": null,
                              "arrivalAt": null,
                              "provenance": {
                                "status": "demo",
                                "sourceIds": [
                                  "demo-journey-scenarios"
                                ],
                                "asOf": null,
                                "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                              }
                            },
                            "provenance": {
                              "status": "demo",
                              "sourceIds": [
                                "demo-journey-scenarios"
                              ],
                              "asOf": null,
                              "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                            }
                          }
                        ],
                        "durationMinutes": 375,
                        "transfers": 0,
                        "emissions": {
                          "status": "demo",
                          "kgCO2ePerTraveler": 18.6,
                          "kgCO2eGroup": 18.6,
                          "coveredDistanceKm": 465,
                          "totalDistanceKm": 465,
                          "comparable": true,
                          "comparisonKey": "carbon-estimation-v1|kgCO2e|kgCO2e/passenger-km|life_cycle|demo|synthetic_test",
                          "methodologyVersion": "carbon-estimation-v1",
                          "assumptions": [],
                          "legs": [
                            {
                              "legId": "outbound-coach-1",
                              "status": "demo",
                              "kgCO2ePerTraveler": 18.6,
                              "kgCO2eGroup": 18.6,
                              "factorId": "demo-coach-factor-v1",
                              "reason": "Calcul de démonstration (provenance des données du trajet)."
                            }
                          ],
                          "factors": [
                            {
                              "id": "demo-coach-factor-v1",
                              "value": 0.04,
                              "unit": "kgCO2e/passenger-km",
                              "mode": "coach",
                              "subtype": null,
                              "geography": "FR",
                              "validFrom": "2026-01-01",
                              "validUntil": null,
                              "scope": "life_cycle",
                              "occupancy": null,
                              "version": "journey-demo-v1",
                              "sourceId": "demo-journey-scenarios",
                              "status": "synthetic_test"
                            }
                          ]
                        },
                        "provenance": {
                          "status": "demo",
                          "sourceIds": [
                            "demo-journey-scenarios"
                          ],
                          "asOf": null,
                          "note": "Scénario synthétique hors ligne; aucun horaire ni offre réelle."
                        },
                        "warnings": [
                          "Durées, attentes, distances et facteurs entièrement synthétiques; aucun horaire réel."
                        ]
                      }
                    ],
                    "warnings": []
                  },
                  "inbound": {
                    "status": "not_requested",
                    "itineraries": [],
                    "warnings": []
                  },
                  "sources": [
                    {
                      "id": "demo-journey-scenarios",
                      "publisher": "Ecotrip — scénarios synthétiques",
                      "url": null,
                      "license": null,
                      "accessedAt": null,
                      "version": "journey-demo-v1",
                      "reuseNotes": "Démonstration et tests uniquement; distances, durées et émissions fictives.",
                      "dataStatus": "demo"
                    }
                  ],
                  "warnings": [
                    "Résultats de démonstration hors ligne; aucune offre réelle ni fallback de fournisseur."
                  ]
                }
              }
            }
          },
          "400": {
            "description": "JSON malformé",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "404": {
            "description": "Ressource inconnue",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "405": {
            "description": "Méthode non autorisée",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "413": {
            "description": "Corps trop grand",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "415": {
            "description": "Type de contenu non pris en charge",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "422": {
            "description": "Critères invalides",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "429": {
            "description": "Limite atteinte; Retry-After en secondes",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "schema": {
                  "type": "integer",
                  "minimum": 1
                },
                "description": "Délai en secondes"
              }
            }
          },
          "503": {
            "description": "Fournisseur indisponible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          },
          "500": {
            "description": "Erreur interne sans détail sensible",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TripRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Mode": {
        "type": "string",
        "enum": [
          "train",
          "coach",
          "walk",
          "public_transport",
          "bicycle",
          "carpool",
          "flight"
        ]
      },
      "Problem": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string",
            "format": "uri-reference"
          },
          "code": {
            "type": "string",
            "enum": [
              "invalid_json",
              "validation_failed",
              "not_found",
              "method_not_allowed",
              "payload_too_large",
              "unsupported_media_type",
              "rate_limited",
              "provider_unavailable",
              "internal_error"
            ]
          },
          "violations": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code"
        ]
      },
      "Source": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
          },
          "publisher": {
            "type": "string"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "nullable": true
          },
          "license": {
            "type": "string",
            "nullable": true
          },
          "accessedAt": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "version": {
            "type": "string",
            "nullable": true
          },
          "reuseNotes": {
            "type": "string"
          },
          "dataStatus": {
            "type": "string",
            "enum": [
              "demo",
              "verified",
              "unverified"
            ]
          }
        },
        "required": [
          "id",
          "publisher",
          "url",
          "license",
          "accessedAt",
          "version",
          "reuseNotes",
          "dataStatus"
        ]
      },
      "Provenance": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "demo",
              "verified",
              "unverified",
              "unknown"
            ]
          },
          "sourceIds": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
            }
          },
          "asOf": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "note": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "sourceIds",
          "asOf",
          "note"
        ]
      },
...
      "TripRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "originId": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
          },
          "destinationId": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
          },
          "departureDate": {
            "type": "string",
            "format": "date"
          },
          "returnDate": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "travelers": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9
          },
          "modes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Mode"
            },
            "minItems": 1,
            "maxItems": 7,
            "uniqueItems": true
          }
        },
        "required": [
          "originId",
          "destinationId",
          "departureDate",
          "travelers",
          "modes"
        ],
        "description": "Dates locales Europe/Paris pour le catalogue MVP. Départ >= aujourd’hui selon horloge serveur injectable, retour >= départ; origine distincte destination. JSON <= 16384 octets. Identifiants inconnus: 422. Une paire connue non couverte: 200/out_of_coverage."
      },
      "Distance": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "km": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "method": {
            "type": "string",
            "enum": [
              "scenario",
              "routed",
              "great_circle",
              "unknown"
            ]
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          }
        },
        "required": [
          "km",
          "method",
          "provenance"
        ]
      },
      "Leg": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
          },
          "mode": {
            "$ref": "#/components/schemas/Mode"
          },
          "subtype": {
            "type": "string",
            "nullable": true
          },
          "originId": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
          },
          "destinationId": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
          },
          "durationMinutes": {
            "type": "integer",
            "minimum": 0
          },
          "waitingMinutes": {
            "type": "integer",
            "minimum": 0
          },
          "distance": {
            "$ref": "#/components/schemas/Distance"
          },
          "schedule": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "departureAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "arrivalAt": {
                "type": "string",
                "format": "date-time",
                "nullable": true
              },
              "provenance": {
                "$ref": "#/components/schemas/Provenance"
              }
            },
            "required": [
              "departureAt",
              "arrivalAt",
              "provenance"
            ]
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          }
        },
        "required": [
          "id",
          "mode",
          "subtype",
          "originId",
          "destinationId",
          "durationMinutes",
          "waitingMinutes",
          "distance",
          "schedule",
          "provenance"
        ],
        "description": "Les durées de scénario ne sont pas des horaires. waitingMinutes est l’attente avant cette étape; pas de double comptage. Aucun horaire déduit de la seule date demandée."
      },
...
      "LegEstimate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "legId": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
          },
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "unavailable",
              "demo"
            ]
          },
          "kgCO2ePerTraveler": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "kgCO2eGroup": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "factorId": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$",
            "nullable": true
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "legId",
          "status",
          "kgCO2ePerTraveler",
          "kgCO2eGroup",
          "factorId",
          "reason"
        ]
      },
      "EmissionEstimate": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "partial",
              "unavailable",
              "demo"
            ]
          },
          "kgCO2ePerTraveler": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "kgCO2eGroup": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "coveredDistanceKm": {
            "type": "number",
            "minimum": 0
          },
          "totalDistanceKm": {
            "type": "number",
            "minimum": 0,
            "nullable": true
          },
          "comparable": {
            "type": "boolean"
          },
          "comparisonKey": {
            "type": "string",
            "nullable": true
          },
          "methodologyVersion": {
            "type": "string"
          },
          "assumptions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "legs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LegEstimate"
            }
          },
          "factors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Factor"
            }
          }
        },
        "required": [
          "status",
          "kgCO2ePerTraveler",
          "kgCO2eGroup",
          "coveredDistanceKm",
          "totalDistanceKm",
          "comparable",
          "comparisonKey",
          "methodologyVersion",
          "assumptions",
          "legs",
          "factors"
        ],
        "description": "Inconnu = null, jamais zéro. Partiel/indisponible non classable carbone. comparisonKey lie unité, périmètre et méthode compatibles; demo reste explicitement synthétique. Les sommes partielles décrivent uniquement les étapes couvertes."
      },
      "Itinerary": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "pattern": "^[a-z0-9][a-z0-9_-]{0,63}$"
          },
          "direction": {
            "type": "string",
            "enum": [
              "outbound",
              "inbound"
            ]
          },
          "requestedDate": {
            "type": "string",
            "format": "date"
          },
          "dataStatus": {
            "type": "string",
            "enum": [
              "demo",
              "real"
            ]
          },
          "legs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Leg"
            },
            "minItems": 1
          },
          "durationMinutes": {
            "type": "integer",
            "minimum": 0
          },
          "transfers": {
            "type": "integer",
            "minimum": 0
          },
          "emissions": {
            "$ref": "#/components/schemas/EmissionEstimate"
          },
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "direction",
          "requestedDate",
          "dataStatus",
          "legs",
          "durationMinutes",
          "transfers",
          "emissions",
          "provenance",
          "warnings"
        ],
        "description": "durationMinutes = somme des durées et attentes. transfers compte les changements de véhicule motorisé; marche de liaison non comptée comme véhicule. Les retours sont recherchés séparément."
      },
      "DirectionResult": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "complete",
              "partial",
              "empty",
              "out_of_coverage",
              "unavailable",
              "not_requested"
            ]
          },
          "itineraries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Itinerary"
            },
            "maxItems": 20
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "status",
          "itineraries",
          "warnings"
        ]
      },
      "JourneysResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "dataMode": {
            "type": "string",
            "enum": [
              "demo",
              "real"
            ]
          },
          "request": {
            "$ref": "#/components/schemas/TripRequest"
          },
          "outbound": {
            "$ref": "#/components/schemas/DirectionResult"
          },
          "inbound": {
            "$ref": "#/components/schemas/DirectionResult"
          },
          "sources": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Source"
            }
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "dataMode",
          "request",
          "outbound",
          "inbound",
          "sources",
          "warnings"
        ],
        "description": "200 empty: couvert mais aucun résultat selon filtres. out_of_coverage: paire hors catalogue. unavailable: échec partiel lorsque autre direction réussit; panne totale = 503. inbound not_requested si pas de retour. Aucun fallback réel vers demo."
      },
...
    }
  }
}